project: rename "yaml" package to "project"
